remove inode update prevention
The mechanism of inode update prevention mechanism was introduced in
the commit 449331b4d0c. The original motivation of this mechanism was
for reducing object leak which can happen during taking
snapshot. However, the object leak problem seems to be solved in the
qemu side commit 20760b16b73. I've been kept the mechanism because it
was useful for hiding incomplete state of VDIs (snapshot is created
but new working VDI isn't created yet) from clients. For simplifing
the design of sheepdog, the problem should be handled in the clients.

The retrying mechanism in the qemu driver is already ongoing:
https://github.com/sheepdog/qemu/tree/reconnect-and-reload

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>

Conflicts:
	dog/vdi.c
	lib/shared/vdi.c
	sheep/request.c